+++ XuanYu.Core/Gizmo/RotateGizmoAxis.cs
@@ -1,0 +1,8 @@
+namespace XuanYu.Core.Gizmo;
+
+public enum RotateGizmoAxis
+{
+    X,
+    Y,
+    Z
+}